Move aggregation factor struct to its own file
